+++ crates/instructions/src/lib.rs
@@ -1,3 +1,26 @@
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
-pub enum BrainInstruction {}
+use frick_operations::{BrainOperation, BrainOperationType};
+
+pub enum BrainInstruction {
+	LoadCell,
+	StoreCell,
+	StartLoop,
+	EndLoop,
+}
+
+pub trait IntoInstructions {
+	fn to_instructions(&self) -> Vec<BrainInstruction>;
+}
+
+impl IntoInstructions for BrainOperation {
+	fn to_instructions(&self) -> Vec<BrainInstruction> {
+		self.ty().to_instructions()
+	}
+}
+
+impl IntoInstructions for BrainOperationType {
+	fn to_instructions(&self) -> Vec<BrainInstruction> {
+		Vec::new()
+	}
+}
