--- crates/new_optimizer/Cargo.toml
@@ -1,17 +1,0 @@
-lints.workspace = true
-
-[package]
-edition.workspace = true
-license.workspace = true
-name = "frick_new_optimizer"
-rust-version.workspace = true
-version.workspace = true
-
-[dependencies]
-frick_operations.workspace = true
-frick_utils = { workspace = true, features = ["insert_or_push", "iter_ext"] }
-serde = { workspace = true, default-features = false, features = [
-    "alloc",
-    "derive"
-] }
-tracing = { workspace = true, features = ["attributes"] }
