+++ crates/old_optimizer/src/inner/passes/loops.rs
